Postavljanje root fragmenta i prilagodba koda
