+++ FOOD2GO/app/src/main/res/layout/content_glavni.xml
@@ -3,16 +3,21 @@
 xmlns:android="http://schemas.android.com/apk/res/android"
 xmlns:app="http://schemas.android.com/apk/res-auto"
 xmlns:tools="http://schemas.android.com/tools"
-android:id="@+id/activity_main"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
 app:layout_behavior="@string/appbar_scrolling_view_behavior"
 tools:context=".GlavniActivity"
 tools:showIn="@layout/app_bar_glavni">
 
-    <FrameLayout
+    <fragment
         android:id="@+id/nav_host_fragment"
-        android:layout_width="420dp"
-        android:layout_height="800dp"></FrameLayout>
+        android:name="androidx.navigation.fragment.NavHostFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:defaultNavHost="true"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:navGraph="@navigation/mobile_navigation" />
 
 </RelativeLayout>
